+++ example/demo.f90
@@ -21,9 +21,12 @@
 ! SOFTWARE.
 !-------------------------------------------------------------------------------
 ! Contributed by vmagnin: 2023-09-26
-! Last modification: gha3mi 2024-01-28, vmagnin 2024-02-22
+! Last modification: gha3mi 2024-01-28, vmagnin 2024-03-01
 !-------------------------------------------------------------------------------
 
+!> This example will create colorbar files for each available colormap and
+!> the corresponding test images. It also demonstrates how you can create your
+!> own colormap defined in an array, or import it from a text file.
 program demo
     use forcolormap, only: Colormap, colormaps_list, wp
     use forcolormap_utils, only: test_colormap
@@ -32,17 +35,24 @@ program demo
     integer :: i
     type(Colormap) :: cmap, custom_cmap
     integer, allocatable :: colors(:,:)
-    ! A discrete colormap with 8 levels, from black to white:
+
+    !> A discrete colormap with 8 levels, by @alozada, resembling the color
+    !> changes in red cabbage (containing Anthocyanins) with pH:
     integer, dimension(0:7, 3) :: my_colormap = reshape( [ &
-          0,     0,     0,   &
-        255,     0,     0,   &
-          0,   255,     0,   &
-          0,     0,   255,   &
-        255,   255,     0,   &
-          0,   255,   255,   &
-        255,     0,   255,   &
-        255,   255,   255 ], &
+        198,    29,    32,   &
+        189,    21,    56,   &
+        171,    82,   150,   &
+        102,    81,   156,   &
+         38,    53,   108,   &
+          5,    65,    40,   &
+        221,   199,    44,   &
+        237,   191,    44 ], &
         shape(my_colormap), order = [2, 1] )
+    !> You can create your own colormap using that array. The name of your
+    !> colormap must conform to the max length defined in colormap_parameters.f90
+    call custom_cmap%create('red_cabbage', 0.0_wp, 2.0_wp, my_colormap)
+    call custom_cmap%colorbar('red_cabbage_colorbar')
+    call test_colormap(custom_cmap, 'red_cabbage_test')
 
     !> We create PPM files (binary encoded by default) for each built-in colormap.
     !> The built-in z=f(x,y) test function is in the [0, 2] range:
@@ -58,11 +68,6 @@ program demo
     ! We change the name for the output test files:
     call cmap%colorbar('cubehelix_customized_colorbar')
     call test_colormap(cmap, 'cubehelix_customized_test')
-
-    ! You can create your own colormap defined in an array:
-    call custom_cmap%create('discrete', 0.0_wp, 2.0_wp, my_colormap)
-    call custom_cmap%colorbar('discrete_colorbar')
-    call test_colormap(custom_cmap, 'discrete_test')
 
     ! Or you can download it from a .txt file:
     call custom_cmap%load("test_map_to_load.txt", 0.0_wp, 2.0_wp)